# YJIT: Decrease SEND_MAX_DEPTH to 5