Since we have the send insn fallback (https://github.com/ruby/ruby/pull/8106), using large `SEND_MAX_DEPTH` no longer contributes to `ratio_in_yjit`.

We tested 4, 5, 8, 10, and 20 as a `SEND_MAX_DEPTH` on SFR, and 5 had the best average latency. It also decreases the max code size by 28%. We also know that something similar happened to Core when we tested 5 and 20 on it.

So this PR changes `SEND_MAX_DEPTH` back to 5, reverting https://github.com/ruby/ruby/pull/7452 and https://github.com/ruby/ruby/pull/7469.