@maximecb 2023-09-07
I think that a 28% reduction in max code size is really great. Could you provide some data for the impact on the performance of headline benchmarks (and also their rss) as well? I would just like to have an idea of the order of magnitude of the impact.