@k0kubun 2023-09-07
Sure.

```
before: ruby 3.3.0dev (2023-09-06T22:26:11Z master 54274b8c65) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T23:54:53Z yjit-depth-5 ac4436f208) +YJIT [x86_64-linux]

--------------  -----------  ----------  ---------  ----------  ----------  ---------  -------------  ------------
bench           before (ms)  stddev (%)  RSS (MiB)  after (ms)  stddev (%)  RSS (MiB)  after 1st itr  before/after
activerecord    30.8         5.3         56.9       30.5        5.3         56.4       1.01           1.01
chunky-png      403.3        0.5         46.4       402.5       0.5         51.1       1.00           1.00
erubi-rails     10.4         32.4        99.0       10.4        26.0        97.0       1.01           1.01
hexapdf         1335.0       2.0         269.0      1335.7      1.1         256.6      1.00           1.00
liquid-c        38.8         0.7         39.7       38.5        0.6         39.5       1.00           1.01
liquid-compile  40.4         0.5         34.8       40.3        0.5         34.6       1.01           1.00
liquid-render   56.7         0.3         38.1       56.3        0.4         37.9       1.00           1.01
lobsters        602.2        4.0         333.7      607.3       4.2         331.0      1.08           0.99
mail            83.4         0.2         50.5       83.6        0.3         50.4       1.02           1.00
psych-load      1246.1       0.2         36.9       1242.6      0.3         36.4       1.00           1.00
railsbench      1130.3       1.3         105.8      1132.2      1.3         102.8      0.99           1.00
ruby-lsp        34.5         24.7        134.3      35.9        26.1        131.7      1.07           0.96
sequel          48.3         1.9         41.0       47.7        1.8         40.8       1.02           1.01
--------------  -----------  ----------  ---------  ----------  ----------  ---------  -------------  ------------
```

1st itr is improved for larger benchmarks like lobsters and ruby-lsp, which would translate to better p90/p99/average latency in production. ruby-lsp seems to get a lower peak performance while its warmup is improved.

RSS is decreased for most benchmarks. chunky-png was the only benchmark that uses more RSS with this change. Its `code_region_size` was actually improved by 8%. `code_region_overhead` happens to be 58 bytes (0.0%) in master while it becomes 25KB (4.7%) in this PR, so it might be related?