@maximecb 2023-09-07
> RSS is decreased for most benchmarks. chunky-png was the only benchmark that uses more RSS with this change.

Benoit Daloze said that RSS is noisy and opened a PR to measure max_rss instead. We may want to switch to using this metric by defaul tin the future. Could also be the overhead as you mentioned, there can be effects that are hard to predict.

I think the results look good. Our benchmarks are probably generally not very polymorphic, so not affected much, which is nice. In the future (YJIT 4.4?) we may want to have a proper inline cache of some sort, which would allow us to have better performance and smaller code size for polymorphic ivar accesses and calls. Either way this is a nice improvement for 3.3.

Worth it to do a similar investigation for getivar max depth.